Refactor `core` module, to break `registers` into a new source file.
