+++ probe-rs/src/architecture/arm/core/armv8a.rs
@@ -1,31 +1,29 @@
 //! Register types and the core interface for armv8-a
 
-use crate::architecture::arm::core::armv8a_debug_regs::*;
-use crate::architecture::arm::memory::adi_v5_memory_interface::ArmProbe;
-use crate::architecture::arm::sequences::ArmDebugSequence;
-use crate::architecture::arm::ArmError;
-use crate::core::{RegisterFile, RegisterValue};
-use crate::error::Error;
-use crate::memory::valid_32bit_address;
-use crate::CoreInterface;
-use crate::CoreStatus;
-use crate::MemoryInterface;
-use crate::RegisterId;
-use crate::{Architecture, CoreInformation, CoreType, InstructionSet};
-use anyhow::Result;
-
-use super::armv8a_core_regs::AARCH64_REGISTER_FILE;
-use super::CortexAState;
-use super::AARCH32_FP_32_REGS;
-
-use super::instructions::aarch64;
-use super::instructions::thumb2::{
-    build_ldr, build_mcr, build_mrc, build_str, build_vmov, build_vmrs,
+use super::{
+    armv8a_core_regs::AARCH64_REGISTER_FILE,
+    instructions::{
+        aarch64,
+        thumb2::{build_ldr, build_mcr, build_mrc, build_str, build_vmov, build_vmrs},
+    },
+    CortexAState, AARCH32_FP_32_REGS,
 };
-
-use std::sync::Arc;
-use std::time::Duration;
-use std::time::Instant;
+use crate::{
+    architecture::arm::{
+        core::armv8a_debug_regs::*, memory::adi_v5_memory_interface::ArmProbe,
+        sequences::ArmDebugSequence, ArmError,
+    },
+    core::{RegisterFile, RegisterId, RegisterValue},
+    error::Error,
+    memory::valid_32bit_address,
+    Architecture, CoreInformation, CoreInterface, CoreStatus, CoreType, InstructionSet,
+    MemoryInterface,
+};
+use anyhow::Result;
+use std::{
+    sync::Arc,
+    time::{Duration, Instant},
+};
 
 /// Errors for the ARMv8-A state machine
 #[derive(thiserror::Error, Debug)]
