+++ probe-rs/src/debug/unit_info.rs
@@ -3,7 +3,7 @@ use super::{
     function_die::FunctionDie, registers, variable::*, DebugError, DebugRegisters, SourceLocation,
     VariableCache,
 };
-use crate::{core::Core, MemoryInterface, RegisterValue};
+use crate::{core::Core, core::RegisterValue, MemoryInterface};
 use gimli::{AttributeValue::Language, Location, UnitOffset};
 use num_traits::Zero;
